Utilities for saving tutkinnot, osaamisalat and tutkintokielet done.
